Modulo 12 - RAG/chat manual/requirements.txt
transformers==4.30.2
huggingface-hub==0.16.4
torch==2.1.0
torchvision==0.16.0
numpy==1.24.3
scikit-learn==1.3.2
PyPDF2==3.0.1
python-docx==1.0.1
beautifulsoup4==4.12.2 